+++ src/View/Application/form/other/Credenciales.java
@@ -44,6 +44,7 @@ import javax.swing.text.BadLocationException;
 import javax.swing.text.PlainDocument;
 import org.json.JSONException;
 import org.json.JSONObject;
+import raven.toast.Notifications;
 
 /**
  *
@@ -76,6 +77,20 @@ public class Credenciales extends javax.swing.JPanel {
                     //FlatAnimatedLafChange.hideSnapshotWithAnimation();
                 });
         }
+        
+        
+        
+         
+        btnReload.setEnabled(false);
+        
+        Timer timer = new Timer(6000, (ActionEvent e) -> { 
+                    
+                     btnReload.setEnabled(true);
+        });
+                timer.setRepeats(false);
+                timer.start();
+        
+        
         lb.putClientProperty(FlatClientProperties.STYLE, ""
                 + "font:$h1.font");
         // Obtén el modelo de la tabla existente
@@ -542,11 +557,37 @@ playAbrir();
         // TODO add your handling code here:
     }//GEN-LAST:event_cbSearchActionPerformed
 
+    private void playError() {
+        String filepath = "src/View/sounds/error.wav";
+
+        PlayMusic(filepath);
+
+    }
+    
     private void btnReloadMouseClicked(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_btnReloadMouseClicked
         // TODO add your handling code here:
-        
-        deleteAllTableRows(table1);
+     
+        if(btnReload.isEnabled() ==false){
+            
+           Notifications.getInstance().clearAll();
+            Notifications.getInstance().show(Notifications.Type.ERROR, Notifications.Location.TOP_CENTER, "Espere un momento");
+                playError();
+        }
+        else{
+            deleteAllTableRows(table1);
         cargarDatos();
+        
+        btnReload.setEnabled(false);
+        
+        Timer timer = new Timer(6000, (ActionEvent e) -> { 
+                    
+                     btnReload.setEnabled(true);
+        });
+                timer.setRepeats(false);
+                timer.start();
+        
+        }
+        
     }//GEN-LAST:event_btnReloadMouseClicked
 
     private void btnReloadActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_btnReloadActionPerformed
