+++ src/View/Application/form/other/RangoDeHoras.java
@@ -37,6 +37,7 @@ import javax.sound.sampled.Clip;
 import javax.swing.JTable;
 import javax.swing.Timer;
 import javax.swing.table.DefaultTableModel;
+import raven.toast.Notifications;
 
 /**
  *
@@ -64,6 +65,20 @@ public class RangoDeHoras extends javax.swing.JPanel {
                     //FlatAnimatedLafChange.hideSnapshotWithAnimation();
                 });
         }
+        
+        
+        btnReload.setEnabled(false);
+        
+        Timer timer = new Timer(1500, (ActionEvent e) -> { 
+                    
+                     btnReload.setEnabled(true);
+        });
+                timer.setRepeats(false);
+                timer.start();
+        
+        
+      
+        
         lb.putClientProperty(FlatClientProperties.STYLE, ""
                 + "font:$h1.font");
         // Obtén el modelo de la tabla existente
@@ -401,11 +416,36 @@ playAbrir();
         // TODO add your handling code here:
     }//GEN-LAST:event_btnAddActionPerformed
 
+    private void playError() {
+        String filepath = "src/View/sounds/error.wav";
+
+        PlayMusic(filepath);
+
+    }
+    
     private void btnReloadMouseClicked(java.awt.event.MouseEvent evt) {//GEN-FIRST:event_btnReloadMouseClicked
         // TODO add your handling code here:
 
-        deleteAllTableRows(table1);
+        if(btnReload.isEnabled() ==false){
+            
+           Notifications.getInstance().clearAll();
+            Notifications.getInstance().show(Notifications.Type.ERROR, Notifications.Location.TOP_CENTER, "Espere un momento");
+                playError();
+        }
+        else{
+           deleteAllTableRows(table1);
         cargarDatosAsync();
+        
+        btnReload.setEnabled(false);
+        
+        Timer timer = new Timer(1500, (ActionEvent e) -> { 
+                    
+                     btnReload.setEnabled(true);
+        });
+                timer.setRepeats(false);
+                timer.start();
+        
+        }
 
     }//GEN-LAST:event_btnReloadMouseClicked
